Move textobjects and surround out of actions.
